iwqi_replication_thesis_package/src/run_full.py
# Placeholder: full run orchestration (load, preprocess, model, evaluate)
# Intentionally minimal to avoid heavy imports in this environment.

import yaml, os, json

def main(config_path: str):
    with open(config_path, "r") as f:
        cfg = yaml.safe_load(f)
    print("Loaded config:", json.dumps(cfg, indent=2)[:400], "...")
    print("This is a placeholder. Implement data loading and training in your environment.")

if __name__ == "__main__":
    import argparse
    ap = argparse.ArgumentParser()
    ap.add_argument("--config", required=True)
    args = ap.parse_args()
    main(args.config)
