iwqi_replication_thesis_package/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
imbalanced-learn==0.12.3
xgboost==2.0.3
lightgbm==4.3.0
matplotlib==3.8.4
seaborn==0.13.2
shap==0.45.1
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.4
nbformat==5.10.4
